+++ game/egoboo_console.c
@@ -379,7 +379,7 @@ SDL_bool egoboo_console_draw( egoboo_console_t * pcon )
 
             // draw the last output line and work backwards
             // this must be iterated with a signed variable or it fails horribly
-            for ( rcnt = (( int )console_line_count ) - 1; rcnt >= 0 && height > 0 ; rcnt-- )
+            for ( rcnt = (( signed )console_line_count ) - 1; rcnt >= 0 && height > 0 ; rcnt-- )
             {
                 size_t len = MIN( 1023, console_line_lengths[rcnt] );
 
@@ -576,7 +576,7 @@ SDL_Event * egoboo_console_handle_events( SDL_Event * pevt )
                 {
                     strncpy( pcon->buffer, pcon->save_buffer[pcon->save_index], SDL_arraysize( pcon->buffer ) );
                     pcon->buffer_carat = strlen( pcon->buffer );
-                    pcon->buffer_carat = (( int )pcon->buffer_carat ) - 1;
+                    pcon->buffer_carat = (( signed )pcon->buffer_carat ) - 1;
                 }
             }
 
@@ -601,7 +601,7 @@ SDL_Event * egoboo_console_handle_events( SDL_Event * pevt )
                 {
                     strncpy( pcon->buffer, pcon->save_buffer[pcon->save_index], EGOBOO_CONSOLE_LENGTH - 1 );
                     pcon->buffer_carat = strlen( pcon->buffer );
-                    pcon->buffer_carat = (( int )pcon->buffer_carat ) - 1;
+                    pcon->buffer_carat = (( signed )pcon->buffer_carat ) - 1;
                 }
             }
